refactor: address PR review comments for typed TaskError migration
- Convert standalone `broadcast_error()` fn to `impl From<SdkError> for TaskError`
  (per @lklimek review), moving logic and tests to `error.rs`
- Add `BroadcastError` variant with user-friendly Display message instead of
  leaking raw SDK text via `TaskError::Generic` (per CodeRabbit review)
- Change `refresh_identity`, `top_up_identity_from_platform_addresses`, and
  `transfer_to_addresses` to return `Result<_, TaskError>` instead of
  `Result<_, String>` (per CodeRabbit review)
- Simplify callers in `run_identity_task` — direct `.await` without
  unnecessary `Ok(self.foo().await?)` wrapping
