+++ src/backend_task/identity/add_key_to_identity.rs
@@ -6,78 +6,17 @@ use crate::model::fee_estimation::PlatformFeeEstimator;
 use crate::model::qualified_identity::PrivateKeyTarget::PrivateKeyOnMainIdentity;
 use crate::model::qualified_identity::QualifiedIdentity;
 use crate::model::qualified_identity::qualified_identity_public_key::QualifiedIdentityPublicKey;
-use dash_sdk::Error as SdkError;
 use dash_sdk::Sdk;
-use dash_sdk::dpp::ProtocolError;
-use dash_sdk::dpp::consensus::ConsensusError;
-use dash_sdk::dpp::consensus::state::state_error::StateError;
 use dash_sdk::dpp::identity::accessors::{IdentityGettersV0, IdentitySettersV0};
 use dash_sdk::dpp::identity::identity_public_key::accessors::v0::{
     IdentityPublicKeyGettersV0, IdentityPublicKeySettersV0,
 };
-use dash_sdk::dpp::platform_value::string_encoding::Encoding;
 use dash_sdk::dpp::prelude::UserFeeIncrease;
 use dash_sdk::dpp::state_transition::identity_update_transition::IdentityUpdateTransition;
 use dash_sdk::dpp::state_transition::identity_update_transition::methods::IdentityUpdateTransitionMethodsV0;
 use dash_sdk::dpp::state_transition::proof_result::StateTransitionProofResult;
 use dash_sdk::platform::transition::broadcast::BroadcastStateTransition;
 use dash_sdk::platform::{Fetch, Identity};
-
-/// Converts a broadcast SDK error into a typed `TaskError`.
-///
-/// Matches known consensus error variants from two SDK error paths
-/// (`StateTransitionBroadcastError` and `Protocol/ConsensusError`),
-/// falling back to `TaskError::Generic` for unrecognised errors.
-/// The original `SdkError` is preserved as a typed field on each variant.
-fn broadcast_error(error: SdkError) -> TaskError {
-    // Classify the error while borrowing; the borrow is dropped before `error`
-    // is moved into the returned variant.
-    enum ConsensusKind {
-        DuplicateKey,
-        DuplicateKeyId,
-        ContractBoundsConflict(String),
-    }
-
-    let kind: Option<ConsensusKind> = {
-        let consensus_error = match &error {
-            SdkError::StateTransitionBroadcastError(broadcast_err) => broadcast_err.cause.as_ref(),
-            SdkError::Protocol(ProtocolError::ConsensusError(ce)) => Some(ce.as_ref()),
-            _ => None,
-        };
-
-        consensus_error.and_then(|ce| match ce {
-            ConsensusError::StateError(StateError::DuplicatedIdentityPublicKeyStateError(_)) => {
-                Some(ConsensusKind::DuplicateKey)
-            }
-            ConsensusError::StateError(StateError::DuplicatedIdentityPublicKeyIdStateError(_)) => {
-                Some(ConsensusKind::DuplicateKeyId)
-            }
-            ConsensusError::StateError(
-                StateError::IdentityPublicKeyAlreadyExistsForUniqueContractBoundsError(e),
-            ) => Some(ConsensusKind::ContractBoundsConflict(
-                e.contract_id().to_string(Encoding::Base58),
-            )),
-            _ => None,
-        })
-    };
-
-    let boxed = Box::new(error);
-    match kind {
-        Some(ConsensusKind::DuplicateKey) => TaskError::DuplicateIdentityPublicKey {
-            source_error: boxed,
-        },
-        Some(ConsensusKind::DuplicateKeyId) => TaskError::DuplicateIdentityPublicKeyId {
-            source_error: boxed,
-        },
-        Some(ConsensusKind::ContractBoundsConflict(contract_id)) => {
-            TaskError::IdentityPublicKeyContractBoundsConflict {
-                contract_id,
-                source_error: boxed,
-            }
-        }
-        None => TaskError::Generic(format!("Broadcasting error: {boxed}")),
-    }
-}
 
 impl AppContext {
     pub(super) async fn add_key_to_identity(
@@ -128,10 +67,7 @@ impl AppContext {
         )
         .map_err(|e| format!("IdentityUpdateTransition error: {}", e))?;
 
-        let result = state_transition
-            .broadcast_and_wait(sdk, None)
-            .await
-            .map_err(broadcast_error)?;
+        let result = state_transition.broadcast_and_wait(sdk, None).await?;
 
         // Log and handle the proof result
         tracing::info!("AddKeyToIdentity proof result: {}", result);
@@ -186,78 +122,5 @@ impl AppContext {
         self.update_local_qualified_identity(&qualified_identity)
             .map(|_| BackendTaskSuccessResult::AddedKeyToIdentity(fee_result))
             .map_err(|e| TaskError::Generic(format!("Database error: {}", e)))
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use dash_sdk::dpp::consensus::state::identity::duplicated_identity_public_key_id_state_error::DuplicatedIdentityPublicKeyIdStateError;
-    use dash_sdk::dpp::consensus::state::identity::duplicated_identity_public_key_state_error::DuplicatedIdentityPublicKeyStateError;
-    use dash_sdk::dpp::consensus::state::identity::identity_public_key_already_exists_for_unique_contract_bounds_error::IdentityPublicKeyAlreadyExistsForUniqueContractBoundsError;
-    use dash_sdk::dpp::identity::Purpose;
-    use dash_sdk::dpp::identifier::Identifier;
-
-    #[test]
-    fn test_duplicate_public_key_error() {
-        let consensus =
-            ConsensusError::from(DuplicatedIdentityPublicKeyStateError::new(vec![1, 2]));
-        let sdk_err = SdkError::from(consensus);
-        let err = broadcast_error(sdk_err);
-        assert!(matches!(err, TaskError::DuplicateIdentityPublicKey { .. }));
-    }
-
-    #[test]
-    fn test_duplicate_public_key_id_error() {
-        let consensus = ConsensusError::from(DuplicatedIdentityPublicKeyIdStateError::new(vec![3]));
-        let sdk_err = SdkError::from(consensus);
-        let err = broadcast_error(sdk_err);
-        assert!(matches!(
-            err,
-            TaskError::DuplicateIdentityPublicKeyId { .. }
-        ));
-    }
-
-    #[test]
-    fn test_contract_bounds_conflict_error() {
-        let contract_id = Identifier::random();
-        let identity_id = Identifier::random();
-        let consensus = ConsensusError::from(
-            IdentityPublicKeyAlreadyExistsForUniqueContractBoundsError::new(
-                identity_id,
-                contract_id,
-                Purpose::AUTHENTICATION,
-                2,
-                1,
-            ),
-        );
-        let sdk_err = SdkError::from(consensus);
-        let err = broadcast_error(sdk_err);
-        let expected_contract_id = contract_id.to_string(Encoding::Base58);
-        assert!(
-            matches!(err, TaskError::IdentityPublicKeyContractBoundsConflict { ref contract_id, .. } if *contract_id == expected_contract_id)
-        );
-    }
-
-    #[test]
-    fn test_broadcast_error_with_cause_duplicate_key() {
-        // Test the StateTransitionBroadcastError path (the actual production path
-        // when Platform returns a structured broadcast error with a consensus cause).
-        let consensus = ConsensusError::from(DuplicatedIdentityPublicKeyStateError::new(vec![1]));
-        let broadcast_err = dash_sdk::error::StateTransitionBroadcastError {
-            code: 40206,
-            message: "duplicate key".to_string(),
-            cause: Some(consensus),
-        };
-        let sdk_err = SdkError::StateTransitionBroadcastError(broadcast_err);
-        let err = broadcast_error(sdk_err);
-        assert!(matches!(err, TaskError::DuplicateIdentityPublicKey { .. }));
-    }
-
-    #[test]
-    fn test_unknown_sdk_error_falls_back() {
-        let sdk_err = SdkError::Generic("connection timeout".to_string());
-        let err = broadcast_error(sdk_err);
-        assert!(matches!(err, TaskError::Generic(ref s) if s.contains("connection timeout")));
     }
 }
