+++ src/backend_task/identity/refresh_identity.rs
@@ -1,4 +1,5 @@
 use crate::app::TaskResult;
+use crate::backend_task::error::TaskError;
 use crate::context::AppContext;
 use crate::model::qualified_identity::{IdentityStatus, QualifiedIdentity};
 use dash_sdk::Sdk;
@@ -14,27 +15,25 @@ impl AppContext {
         sdk: &Sdk,
         qualified_identity: QualifiedIdentity,
         sender: crate::utils::egui_mpsc::SenderAsync<TaskResult>,
-    ) -> Result<BackendTaskSuccessResult, String> {
+    ) -> Result<BackendTaskSuccessResult, TaskError> {
         let refreshed_identity_id = qualified_identity.identity.id();
         // Fetch the latest state of the identity from Platform
         let maybe_refreshed_identity = Identity::fetch_by_identifier(sdk, refreshed_identity_id)
             .await
-            .map_err(|e| e.to_string())?;
+            .map_err(|e| TaskError::Generic(format!("Failed to fetch identity: {}", e)))?;
 
         // Get local identities
-        let mut local_qualified_identities = self
-            .load_local_qualified_identities()
-            .map_err(|e| e.to_string())?;
+        let mut local_qualified_identities = self.load_local_qualified_identities()?;
 
         // Find the local identity to update
         let outdated_identity_index = local_qualified_identities
             .iter()
             .position(|qi| qi.identity.id() == refreshed_identity_id)
             .ok_or_else(|| {
-                format!(
+                TaskError::Generic(format!(
                     "Identity with id {} not found in local identities",
                     refreshed_identity_id.to_string(Encoding::Base58)
-                )
+                ))
             })?;
 
         // Remove the outdated identity from local state
@@ -50,7 +49,6 @@ impl AppContext {
                     .update(IdentityStatus::Active);
             }
             None => {
-                // it is not found and the status allows refresh, update status to NotFound
                 qualified_identity_to_update
                     .status
                     .update(IdentityStatus::NotFound);
@@ -58,14 +56,13 @@ impl AppContext {
         }
 
         // Insert the updated identity into local state
-        self.update_local_qualified_identity(&qualified_identity_to_update)
-            .map_err(|e| e.to_string())?;
+        self.update_local_qualified_identity(&qualified_identity_to_update)?;
 
         // Send refresh message to refresh the Identities Screen
         sender
             .send(TaskResult::Refresh)
             .await
-            .map_err(|e| e.to_string())?;
+            .map_err(|e| TaskError::Generic(e.to_string()))?;
 
         Ok(BackendTaskSuccessResult::RefreshedIdentity(
             qualified_identity,
